Create Dump20250522 (1).sql
Presentation of the database (applicationdb) in MySQL Workbench consisting of 3 tables (teachers, courses, teacher_courses). 

The teachers table consists of 3 columns (id, firstname, lastname). 

The courses table has two columns: id (primary key, auto-increment for unique identification) and course_name. 

The teacher_courses table has teacher_id and course_id as primary keys, ensuring that each teacher is exclusively linked to the corresponding course. 
In the teacher_courses table, the fields teacher_id and course_id are defined as foreign keys: teacher_id refers to the id of the teachers table, ensuring the existence of the teacher, while course_id refers to the id of the courses table, linking the course to the corresponding teacher.