Application for managing teachers & courses_demo
Presentation through a short demo video of the teacher management application and its connection to their teaching subject in a graphical environment. 

Specifically, in the video, incorrect credentials are entered, and the corresponding warning pop-up message is displayed for the application user, as well as the correct use of IDs and the appearance of the next window to search for a teacher or information about the program by clicking the version info button. 

In the registration window, a new teacher is registered in the school's secretariat database, and in the search window, a teacher is searched by retrieving search results from the database.

After searching for a professor in the results window, a record of a professor is modified with a pop-up notification of successful modification, as well as a corresponding pop-up in the case of deleting a professor from the database, while simultaneously demonstrating the use of the navigation buttons for various records in the results window.

In the demo, the presentation of edge cases or incorrect use of the program by the user is conducted, as well as the expected response of the software (the program) to these edge cases.

More specifically, in the search window, the user does not enter data in the Last Name field, and by clicking the Search button, a pop-up appears guiding the application user to complete the required fields. At the same time, the search for a teacher is being conducted using inappropriate characters and the appearance of the corresponding pop-up window for correction by the user (e.g., Zach41). Additionally, in the search window, the input of the surname is done with all letters being lowercase or with all letters being uppercase (e.g., pavli and PAVLI), and a pop-up appears with the automatic correction of the surname and its display in the next results window.

In the registration window, the user does not fill in the Name and Last Name fields, resulting in the corresponding pop-up notification to the user about filling in all the required fields for completing a teacher's registration in the database, while at the same time, a corresponding pop-up appears indicating the inability to register a teacher in case the user enters a number in one. At the same time, the registration of a professor with 2 names is taking place and a pop-up of successful entry into the database appears, while simultaneously the attempt to register a professor with the same first and last name by the user displays a pop-up notification for an already existing record to prevent the insertion of duplicate entries.

In the results window, deletion and updating are performed with empty fields, resulting in the corresponding pop-up notifications for the user appearing, indicating the inability to execute such actions without data in the respective fields. Simultaneously, in the results window, if the user has a record in front of them and does not modify at least one of the two fields (either the first name or the last name), a pop-up appears notifying the user of the lack of changes, and by pressing the OK button, the results window is reactivated, keeping the record unchanged.