Application Database.sql
This application connects through the Workbench to the database to send SQL queries. The table teachers consists of a column id for each teacher as the Primary Key, name, and lastname.

More specifically, in the implementation of the application, a database (applicationdb) was created consisting of 3 tables (teachers, courses, teacher_courses).

The teachers table consists of 3 columns (id, firstname, lastname).

The courses table has two columns: id (primary key, auto-increment for unique identification) and course_name.

The teacher_courses table has the columns teacher_id and course_id as primary keys, ensuring that each teacher is exclusively linked to the corresponding course. In the teacher_courses table, the fields teacher_id and course_id are defined as foreign keys: teacher_id refers to the id of the teachers table, ensuring the existence of the teacher, while course_id refers to the id of the courses table, linking the course with the corresponding teacher.